Update builder.rb
When I run RSpec test and I got exception before the test runs, then I can't get attachments of my run. 
I change the code to fix this: if the test is not started, we will add attachments to the suite instead of the test.
This change works on my local.